refactor: 변수명 변경, 파일 분리
- constants -> product, chargeMoney, errorMessage, routes 로 분리
- dom -> checkErrorMessage, domHelper, snackBar로 분리
- utils -> validation(checkChargeMoney, checkProduct)로 분리

- coinInput -> chargeMoneyInput
- totalCoin -> totalChargeMoney
- chargeCoinSection -> chargeMoneySection
