+++ index.html
@@ -100,16 +100,16 @@
           </table>
         </section>
       </section>
-      <section class="charge-coin-section hide">
+      <section class="charge-money-section hide">
         <h2 hidden>잔돈 충전</h2>
         <section class="charge-form-section">
           <h3 hidden>자판기가 보유할 금액</h3>
           <form class="charge-form-section__form">
-            <label for="charge-coin">자판기 보유할 금액을 입력해주세요.</label>
-            <div class="charge-form-section__coin-input-wrapper">
+            <label for="charge-money">자판기 보유할 금액을 입력해주세요.</label>
+            <div class="charge-form-section__charge-money-input-wrapper">
               <input
-                class="input charge-form-section__coin-input"
-                name="charge-coin"
+                class="input charge-form-section__charge-money-input"
+                name="charge-money"
                 placeholder="금액"
                 type="number"
                 step="10"
@@ -125,7 +125,7 @@
           </form>
           <p>
             현재 보유 금액:
-            <span class="charge-form-section__total-coin">0</span>원
+            <span class="charge-form-section__total-charge-money">0</span>원
           </p>
         </section>
         <section class="coin-quantity-section">
