+++ src/ts/constants/routes.ts
@@ -1,0 +1,7 @@
+import { Routes } from '../types/constants';
+
+export const ROUTES: Routes = {
+  HOME: '/',
+  PRODUCTS: '/products',
+  COINS: '/coins',
+};
