+++ src/ts/types/vendingMachineChargeMoneyManager.ts
